masHalos computes density

+++ src/masHalos.h
@@ -1,1 +1,1 @@
-void masHalos(int ng, int nH, double boxsize, float *pos, float *density, int *Nx);
+void masHalos(int ng, int nH, double boxsize, float *pos, float *density, float *Nx);
